backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose==3.3.0
passlib==1.7.4
websocket-client==1.8.0
requests==2.32.3
fake-useragent==2.0.3
